feat(rpg): открытый мир, кланы и строительство базы
Полная реализация RPG-системы из 3 фаз:

Фаза 1 — Открытый мир:
- Бесконечная процедурная карта с чанками 32x32 тайлов
- WorldEngine с тик-лупом 20 Гц, загрузка/выгрузка чанков
- Зомби с AI (поиск игрока → атака), сохранение в БД
- Инвентарь (металл, дерево, еда, патроны, аптечки)
- Сбор ресурсов, лут с зомби, респавн с safe zone
- Phaser-рендерер с камерой, спрайтами и HUD

Фаза 2 — Кланы:
- REST API: создание, вступление, выход, кик, повышение
- Депозит ресурсов из инвентаря в клан
- Фронтенд экран клана с участниками и ресурсами

Фаза 3 — Строительство базы:
- Сетка 16x16, типы зданий (стены, турели, шахты, фермы)
- REST API: размещение, сбор ресурсов, снос
- Турели автоматически стреляют по зомби
- Canvas-рендерер базы с режимом строительства

Дополнительно:
- Детальные SVG/PNG спрайты зомби и игрока
- DEV-режим для тестирования без Telegram
- Клавиатурное управление (WASD + стрелки)

+++ backend/app/models.py
@@ -193,7 +193,7 @@ class MapChunk(Base):
     spawn_points = Column(JSON, nullable=True)
 
     # Generation seed (for reproducibility)
-    seed = Column(Integer, nullable=False)
+    seed = Column(BigInteger, nullable=False)
 
     generated_at = Column(DateTime, server_default=func.now())
